nacl: Add custom thread implementation
Similar to Android, NaCl has a pthread implementation without support
for thread cancellation.

We reuse android/thread in order to minimize new threading code
but we complement that with a custom implementation of futexes,
which falls back to pthread mutexes and condition variables.
Actual futexes are not available to regular NaCl modules.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>