WIP: slightly improve rate change
For now the main issue is that the audio output thread is only "timed" when its buffer is full. It kinda work with pulse since it fills the buffer with zeroes and ensure buffer fullness.
For others we need a way to wait but not too much else it will create small audio gaps.
In the end I think we need an audio output rework to:

* Allow feeding data without playing to have a sync preroll (with the dejitter delay)
* Better way to report time. Polling is ok but I feel like time_get fall a litlle short of what it could achieve.
* A proper way to asserve the audio decoding thread while maintaining buffer fullness.

So the next challenge is how to determine when we start the synchro.