FIXUP: fix preroll with !pulse aouts
The idea here is a play a silence corresponding to the difference between the
first system pts and now. This will be used for aouts that can't start corked,
so every aout modules except PulseAudio. To reach this new code path, aout
modules need to return a valid value from the first pf_get_time callback
(already the case of main modules).